# perf: optimize trace return and revert values