### What I did

fixes: #80 
Fixes: APE-1799

Utilize the onlytopLevelcall config for ape-alchemy to make revert reason and return data grabbing much faster in ape-alchemy:

In [1]: txn_hash = "0x36144f609e0fc7afd3cc570d6a54582091642a44c5223a5ad59aa20008dd9577"
In [2]: receipt = chain.history[txn_hash]
In [3]: %time receipt.revert_message

Before: 1.52 s, 951 ms, 907 ms, 919 ms
After: 158 ms, 219 ms, 179 ms, 172 ms

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)